refactor(sandbox): coarsen backend seam to execute()
Replace the Phase-1 fine-grained createWorkspace/ensureCacheStore/launch
methods on ExecutionBackend with a single execute(cfg, req, opts) →
ExecuteResponse. The local-workspace orchestration (stage → run → phase
parse + live tail → harvest → response assembly) moves verbatim out of
spawn.ts into a shared runLocalWorkspaceExecution helper that
DockerBackend.execute delegates to. spawn.ts is now a thin dispatcher:
request validation + the in-flight registry + cancelExecution.

The fine-grained trio is now the DockerBackend-internal LocalWorkspaceRuntime
contract (still satisfied by both backends). The exec-tar KubernetesBackend
delegates to the same helper for now; the exec-free HTTP-only rewrite that
needs the coarse seam lands in the following commits.

Behavior is byte-identical — the existing spawn-staging / spawn-prior-outputs
tests (136 pass) plus typecheck/lint/format are the parity proof.
